+++ commands.js
@@ -5,6 +5,7 @@ import { resetJerryWorkshopTracker } from "./features/overlays/jerryWorkshopTrac
 import { resetWormMembraneProfitTracker } from "./features/overlays/wormMembraneProfitTracker";
 import { resetMagmaCoreProfitTracker } from "./features/overlays/magmaCoreProfitTracker";
 import { resetFishingProfitTracker } from "./features/overlays/fishingProfitTracker";
+import { resetDropNumbers } from "./features/chat/messageOnDrop";
 
 register("command", (...args) => {
     settings.openGUI();
@@ -44,6 +45,9 @@ register("command", (...args) => {
 register("command", (...args) => {
     const isConfirmed = args[0] && args[0] === "noconfirm";
     resetFishingProfitTracker(!!isConfirmed);
+    if (isConfirmed) {
+        resetDropNumbers();
+    }
     return;
 }).setName("feeshResetProfitTracker");