+++ constants/drops.js
@@ -1,6 +1,8 @@
 export const BABY_YETI_PET = "Baby Yeti";
 export const FLYING_FISH_PET = "Flying Fish";
 export const MEGALODON_PET = "Megalodon";
+export const SQUID_PET = "Squid";
+export const GUARDIAN_PET = "Guardian";
 export const LUCKY_CLOVER_CORE = "Lucky Clover Core";
 export const DEEP_SEA_ORB = "Deep Sea Orb";
 export const RADIOACTIVE_VIAL = "Radioactive Vial";